Thread ID Cleanup, main branch (2025.01.09.) (#810)
* Harmonized the different thread_id classes with each other.

Made all of them into "private headers", and added automated tests that they
would fulfill the appropriate concept.

* Synchronized how thread IDs would be used in the device functions.

While also cleaning up the includes of the files a bit.

* Call device functions consistently from CUDA.

* Call device functions consistently from SYCL.

* Implement (some of) Stephen's suggestions.

+++ device/common/include/traccc/clusterization/device/impl/ccl_kernel.ipp
@@ -1,6 +1,6 @@
 /** TRACCC library, part of the ACTS project (R&D line)
  *
- * (c) 2022-2024 CERN for the benefit of the ACTS project
+ * (c) 2022-2025 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
@@ -254,8 +254,7 @@ TRACCC_DEVICE inline void ccl_kernel(
      */
     if (thread_id.getLocalThreadIdX() == 0) {
         unsigned int start =
-            static_cast<unsigned int>(thread_id.getBlockIdX()) *
-            cfg.target_partition_size();
+            thread_id.getBlockIdX() * cfg.target_partition_size();
         assert(start < num_cells);
         unsigned int end =
             std::min(num_cells, start + cfg.target_partition_size());
